// src/app/core/models/dto/IUsuarioDto.ts
export interface IUsuarioDto {
    usuarioNombre?: string | null;
    password?: string | null;
    status?: number | 1;
    clienteId?: string | null;
    usuarioId?: string | null;
    nombre?: string | null;
    apellidos?: string | null;
    direccion?: string | null;
    fecha?: Date | null;
  }
  